convert images dtype to float before padding

+++ utils.py
@@ -13,14 +13,14 @@ def pad_unfoldNd(images: torch.Tensor, kernel_size=3, pad_mode='circular') -> to
     pad = tuple(k // 2 for k in kernel_size for _ in range(dims)) #calculate padding needed based on kernel_size
 
     if pad_mode == "circular":
-        padded_images = f.pad(images, pad, pad_mode) #if "pad_mode"!=list, pad dimensions simultaneously
+        padded_images = f.pad(images.to(dtype=torch.float32), pad, pad_mode) #if "pad_mode"!=list, pad dimensions simultaneously
     elif pad_mode == 'constant':
-        padded_images = f.pad(images, pad, pad_mode, value=-1) 
+        padded_images = f.pad(images.to(dtype=torch.float32), pad, pad_mode, value=-1) 
     
 
     unfold_images = unfoldNd(padded_images.unsqueeze(1), kernel_size=kernel_size) #shape = [N, product(kernel_size), dim1*dim2*dim3]
 
-    return unfold_images
+    return unfold_images.to(dtype=images.dtype)
 
 def ndifferent_neighbors(unfold_images: torch.Tensor) -> torch.Tensor:  
     #ims_unfold - torch tensor of shape = [N, product(kernel_size), dim1*dim2] from [N, 1, dim1, dim2] using "torch.nn.Unfold" object
